// app/models/invoice.ts
export class Invoice {

    id: number;
    reverseCharge: string;
    invoiceDate: string;
    state: string;
    stateCode: number;
    transportationMode: string;
    vehicleNumber: string;
    dateOfSupply: string;
    placeOfSupply: string;
    customerName: string;
    customerAddress: string;
    consigneeName: string;
    consigneeAddress: string;
    hsnCode: number;
}